Bucket sort
Bucket sort runs in linear time when the input is drawn from a uniform distribution.Like counting sort, bucket sort is fast because it assumes something about the input.Whereas counting sort assumes that the input consists of integers in a small range,bucket sort assumes that the input is generated by a random process that distributes elements uniformly over the interval  [0,1).The idea of bucket sort is to divide the interval [0,1)into n equal-sized sub in-tervals, or buckets, and then distribute input numbers into the buckets. Since the inputs are uniformly distributed over [0,1), we don’t expect many numbers to fall into each bucket.  To produce the output, we simply sort the numbers in each bucket and then go through the buckets in order, listing the elements in each.